Collected all remaining data
Pulled 'image_predictions.tsv' using Requests, used Twitter's API to
query the information for all tweets included in the archive, then
saved that information in JSON format in a txt file with each tweet
representing a new line. Read that file line by line into a list, and
then converted that list into a pandas DataFrame to collect the
retweet_count and favorite_count information for each tweet, and then
merged that with the archive information provided from Udacity to create
a new tweet_data dataframe with all of that information.

Gathering of the information is complete, so next steps are to assess
the data and find the issues to resolve.