Completed cleaning stage, exported clean dataframe
Finished the initial run through of cleaning the data, including
8 quality issues and 3 tidiness issues. I think I need to replace
one of the quality issues, and then go through and double check
everything. At the very end, exported the updated dataframe to a csv.

I've been jumping around with the coding a lot, so in a later draft
I'll need to resort and rerun the whole notebook to make sure everything
still works.